+++ kfinance/kfinance.py
@@ -1644,9 +1644,9 @@ class Client:
             for tool_cls in ALL_TOOLS:
                 tool = tool_cls(kfinance_client=self)  # type: ignore[call-arg]
                 if (
-                    tool.required_permission is None
+                    tool.accepted_permissions is None
                     # if one or more of the required permission for a tool is a permission the user has
-                    or tool.required_permission.intersection(
+                    or tool.accepted_permissions.intersection(
                         self.kfinance_api_client.user_permissions
                     )
                 ):