+++ kfinance/tests/test_client.py
@@ -13,6 +13,7 @@ from kfinance.tool_calling import (
     GetNextEarnings,
     GetTranscript,
 )
+from kfinance.tool_calling.shared_models import KfinanceTool
 
 
 class TestLangchainTools:
@@ -57,21 +58,31 @@ class TestLangchainTools:
         # User should not have access to functions that require statement permissions
         assert GetFinancialStatementFromIdentifier not in tool_classes
 
-    def test_permission_set_handling(self, mock_client: Client):
+    @pytest.mark.parametrize(
+        "user_permission, expected_tool",
+        [
+            pytest.param(
+                Permission.TranscriptsPermission, GetTranscript, id="Transcript Permissions"
+            ),
+            pytest.param(Permission.EarningsPermission, GetEarnings, id="Earnings Permissions"),
+        ],
+    )
+    def test_permission_set_handling(
+        self, user_permission: Permission, expected_tool: KfinanceTool, mock_client: Client
+    ):
         """
         GIVEN a user with a permission that is in a set of required permission for a tool
         WHEN we filter tools by permissions
         THEN we successfully return the correct tools that either don't require permissions or tools that the user
             specifically has access to.
         """
-
-        mock_client.kfinance_api_client._user_permissions = {Permission.TranscriptsPermission}  # noqa: SLF001
+        mock_client.kfinance_api_client._user_permissions = {user_permission}  # noqa: SLF001
         tool_classes = [type(t) for t in mock_client.langchain_tools]
         # User should have access to GetEarnings, GetNextEarnings, GetLatestEarnings, GetTranscript
         assert GetEarnings in tool_classes
         assert GetNextEarnings in tool_classes
         assert GetLatestEarnings in tool_classes
-        assert GetTranscript in tool_classes
+        assert expected_tool in tool_classes
         # User should have access to functions that don't require permissions
         assert GetLatest in tool_classes
         # User should not have access to functions that require statement permissions