+++ kfinance/tool_calling/get_cusip_from_ticker.py
@@ -14,7 +14,7 @@ class GetCusipFromTicker(KfinanceTool):
     name: str = "get_cusip_from_ticker"
     description: str = "Get the CUSIP associated with a ticker."
     args_schema: Type[BaseModel] = GetCusipFromTickerArgs
-    required_permission: set[Permission] | None = {Permission.IDPermission}
+    accepted_permissions: set[Permission] | None = {Permission.IDPermission}
 
     def _run(self, ticker_str: str) -> str:
         return self.kfinance_client.ticker(ticker_str).cusip