+++ kfinance/tool_calling/get_info_from_identifier.py
@@ -10,7 +10,7 @@ class GetInfoFromIdentifier(KfinanceTool):
     name: str = "get_info_from_identifier"
     description: str = "Get the information associated with an identifier. Info includes company name, status, type, simple industry, number of employees (if available), founding date, webpage, HQ address, HQ city, HQ zip code, HQ state, HQ country, and HQ country iso code."
     args_schema: Type[BaseModel] = ToolArgsWithIdentifier
-    required_permission: set[Permission] | None = None
+    accepted_permissions: set[Permission] | None = None
 
     def _run(self, identifier: str) -> str:
         return str(self.kfinance_client.ticker(identifier).info)