+++ kfinance/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## v2.6.4
-- Add transcript permission and update required_permissions type to set[Permission]
+- Add transcript permission and update accepted_permissions type to set[Permission]
 
 ## v2.6.3
 - Bump urllib3 to 2.5 to  address CVEs